requirements.txt
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-appconf==1.0.3
django-countries==5.3.3
django-crispy-forms==1.7.2
django-imagekit==4.0.2
django-taggit==1.1.0
httpie==1.0.2
idna==2.8
oauthlib==3.0.2
pilkit==2.0
Pillow==6.1.0
Pygments==2.4.2
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
